yolov5/requirements.txt

# Usage: pip install -r requirements.txt
# -f https://download.pytorch.org/whl/cu116
# torch==1.13.0+cu116
# torchvision==0.14.0+cu116
# torchaudio==0.13.0

gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.22.2
opencv-python>=4.1.1
Pillow>=10.0.1
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation

torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.147
pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1
albumentations>=1.0.3
pycocotools>=2.0.6
